+++ rheia/CASES/PYTEST_LOGNORMAL_EQUIV/case_description.py
@@ -1,0 +1,8 @@
+
+def evaluate(x_in, params=[]):
+    _, sample = x_in
+    u_1 = sample['u_1']
+    u_2 = sample['u_2']
+    x = sample['x']
+    output = 2. + 0.4 * u_1 - 0.2 * u_2 + 0.3 * x + 0.01 * x * u_1
+    return (output,)